kafka-consumer-group-python/requirements.txt
kafka-python==2.0.2
kafka-python-ng
requests==2.31.0
sseclient-py==1.7.2